near-operation-file-preset to colocate codegen generated files
