+++ packages/client/.eslintignore
@@ -1,3 +1,3 @@
 dist
 *.cjs
-src/graphql/generated.ts
+*.generated.ts
