+++ packages/client/package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^1.21.7",
+    "@graphql-codegen/near-operation-file-preset": "^2.5.0",
     "@graphql-codegen/typescript": "^1.23.0",
     "@graphql-codegen/typescript-graphql-request": "^4.5.8",
     "@graphql-codegen/typescript-operations": "^1.18.4",
