+++ packages/client/src/client.ts
@@ -1,26 +1,24 @@
 import { GraphQLClient } from 'graphql-request';
 
-import { Auth } from './auth';
-import { EnvironmentConfig } from './environments';
-import { getSdk, Sdk } from './graphql/generated';
-import { Profile } from './profile';
-import { Publication } from './publication';
+import { Auth } from './auth/auth';
+import { EnvironmentConfig } from './consts/environments';
+import { Profile } from './profile/profile';
+import { Publication } from './publication/publication';
 
 export type LensClientConfig = {
   environment: EnvironmentConfig;
 };
 
 export class LensClient {
-  private readonly sdk: Sdk;
   readonly auth: Auth;
   readonly profile: Profile;
   readonly publication: Publication;
 
   constructor(config: LensClientConfig) {
     const client = new GraphQLClient(config.environment.backend);
-    this.sdk = getSdk(client);
-    this.auth = new Auth(this.sdk);
-    this.profile = new Profile(this.sdk);
-    this.publication = new Publication(this.sdk);
+
+    this.auth = new Auth(client);
+    this.profile = new Profile(client);
+    this.publication = new Publication(client);
   }
 }
