+++ packages/client/src/index.ts
@@ -1,1 +1,2 @@
 export * from './client';
+export * from './consts/environments';
