+++ packages/client/src/profile/graphql/mutations.graphql
@@ -1,0 +1,54 @@
+mutation CreateFollowTypedData($request: FollowRequest!, $options: TypedDataOptions) {
+  result: createFollowTypedData(request: $request, options: $options) {
+    id
+    expiresAt
+    typedData {
+      types {
+        FollowWithSig {
+          name
+          type
+        }
+      }
+      domain {
+        ...EIP712TypedDataDomain
+      }
+      value {
+        nonce
+        deadline
+        profileIds
+        datas
+      }
+    }
+  }
+}
+
+mutation CreateUnfollowTypedData($request: UnfollowRequest!) {
+  result: createUnfollowTypedData(request: $request) {
+    id
+    expiresAt
+    typedData {
+      types {
+        BurnWithSig {
+          name
+          type
+        }
+      }
+      domain {
+        ...EIP712TypedDataDomain
+      }
+      value {
+        nonce
+        deadline
+        tokenId
+      }
+    }
+  }
+}
+
+fragment EIP712TypedDataDomain on EIP712TypedDataDomain {
+  __typename
+  name
+  chainId
+  version
+  verifyingContract
+}
