+++ packages/client/src/publication/graphql/publication.graphql
